Rename file now it is more generic components folder
