+++ 2024-frontend/src/components/WasteMeasurementDetail.vue
